+++ lib/gds_api/test_helpers/content_store.rb
@@ -1,11 +1,9 @@
 require 'gds_api/test_helpers/json_client_helper'
-require 'gds_api/test_helpers/content_item_helpers'
 require 'json'
 
 module GdsApi
   module TestHelpers
     module ContentStore
-      include ContentItemHelpers
 
       CONTENT_STORE_ENDPOINT = Plek.current.find('content-store')
 
@@ -22,6 +20,28 @@ module GdsApi
 
       def content_store_isnt_available
         stub_request(:any, /#{CONTENT_STORE_ENDPOINT}\/.*/).to_return(:status => 503)
+      end
+
+      def content_item_for_base_path(base_path)
+        {
+          "title" => titleize_base_path(base_path),
+          "description" => "Description for #{base_path}",
+          "format" => "guide",
+          "need_ids" => ["100001"],
+          "public_updated_at" => "2014-05-06T12:01:00+00:00",
+          "base_path" => base_path,
+          "details" => {
+            "body" => "Some content for #{base_path}",
+          }
+        }
+      end
+
+      def titleize_base_path(base_path, options = {})
+        if options[:title_case]
+          base_path.gsub("-", " ").gsub(/\b./) {|m| m.upcase }
+        else
+          base_path.gsub(%r{[-/]}, " ").strip.capitalize
+        end
       end
     end
   end
