+++ lib/gds_api/test_helpers/publishing_api.rb
@@ -1,26 +1,30 @@
 require 'gds_api/test_helpers/json_client_helper'
-require 'gds_api/test_helpers/content_item_helpers'
 require 'gds_api/test_helpers/intent_helpers'
 require 'json'
 
 module GdsApi
   module TestHelpers
     module PublishingApi
-      include ContentItemHelpers
       include IntentHelpers
 
       PUBLISHING_API_ENDPOINT = Plek.current.find('publishing-api')
 
-      def stub_publishing_api_put_item(base_path, body = content_item_for_base_path(base_path))
+      def stub_publishing_api_put_item(base_path, body = content_item_for_base_path_in_a_publish_request(base_path))
+        raise ArgumentError, "Strings are no longer supported for response bodies" if body.is_a?(String)
+
         url = PUBLISHING_API_ENDPOINT + "/content" + base_path
-        body = body.to_json unless body.is_a?(String)
-        stub_request(:put, url).with(body: body).to_return(status: 201, body: body, headers: {})
+        response_body = body.dup
+        response_body["base_path"] = base_path
+        stub_request(:put, url).with(body: body).to_return(status: 201, body: response_body.to_json, headers: {})
       end
 
       def stub_publishing_api_put_intent(base_path, body = intent_for_base_path(base_path))
+        raise ArgumentError, "Strings are no longer supported for response bodies" if body.is_a?(String)
+
         url = PUBLISHING_API_ENDPOINT + "/publish-intent" + base_path
-        body = body.to_json unless body.is_a?(String)
-        stub_request(:put, url).with(body: body).to_return(status: 201, body: body, headers: {})
+        response_body = body.dup
+        response_body["base_path"] = base_path
+        stub_request(:put, url).with(body: body).to_return(status: 201, body: response_body.to_json, headers: {})
       end
 
       def stub_publishing_api_destroy_intent(base_path)
@@ -62,6 +66,27 @@ module GdsApi
 
       def publishing_api_isnt_available
         stub_request(:any, /#{PUBLISHING_API_ENDPOINT}\/.*/).to_return(:status => 503)
+      end
+
+      def content_item_for_base_path_in_a_publish_request(base_path)
+        {
+          "title" => titleize_base_path(base_path),
+          "description" => "Description for #{base_path}",
+          "format" => "guide",
+          "need_ids" => ["100001"],
+          "public_updated_at" => "2014-05-06T12:01:00+00:00",
+          "details" => {
+            "body" => "Some content for #{base_path}",
+          }
+        }
+      end
+
+      def titleize_base_path(base_path, options = {})
+        if options[:title_case]
+          base_path.gsub("-", " ").gsub(/\b./) {|m| m.upcase }
+        else
+          base_path.gsub(%r{[-/]}, " ").strip.capitalize
+        end
       end
     end
   end
